+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [2.0.0+1] - 2021.03.22
+
+- Add Backdrop Bars, option to show/hide.
+- Add description widget, located below the bars. Option to always show or by tapping the bars.
+- Add config for tooltip size thanks to @[daniloadorno](https://github.com/daniloadorno)
+
 ## [2.0.0] - 2021.03.17
 
 - Support Null-safety